requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2
django-composite-foreignkey==1.1.0
django-jquery==3.1.0
django-selectize==1.2
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
numpy==1.20.2
pandas==1.2.4
psycopg2==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
